# Conflicts:
#	main.cpp
#	src/Restaurant.cpp

+++ 203458674-308429612/src/Table.cpp
@@ -60,7 +60,7 @@ void Table::order(const std::vector<Dish> &menu) {
             }
         }
     }
-   for(int i=0;i<thisOrder.size();i++){
+   for(int i=0;i<(int)thisOrder.size();i++){
        std::cout<<getCustomer(thisOrder[i].first)->getName()<<  " ordered " << thisOrder[i].second.getName()<<std::endl;
    }
 }
